+++ libnymea-app/types/ruleactions.cpp
@@ -33,7 +33,7 @@ RuleActions::RuleActions(QObject *parent) : QAbstractListModel(parent)
 int RuleActions::rowCount(const QModelIndex &parent) const
 {
     Q_UNUSED(parent)
-    return m_list.count();
+    return static_cast<int>(m_list.count());
 }
 
 QVariant RuleActions::data(const QModelIndex &index, int role) const
@@ -46,7 +46,7 @@ QVariant RuleActions::data(const QModelIndex &index, int role) const
 void RuleActions::addRuleAction(RuleAction *ruleAction)
 {
     ruleAction->setParent(this);
-    beginInsertRows(QModelIndex(), m_list.count(), m_list.count());
+    beginInsertRows(QModelIndex(), static_cast<int>(m_list.count()), static_cast<int>(m_list.count()));
     m_list.append(ruleAction);
     endInsertRows();
     emit countChanged();
